refactor: Define django application for core features
