+++ server/apps/core/admin.py
