+++ server/apps/core/shared_constants.py
