+++ server/apps/core/templates/core/index.html
