--- server/apps/learning/models/__init__.py
@@ -1,2 +1,0 @@
-from .common_models import Course, VOSOrganization
-from .person import Person
